+++ Clean.ps1
@@ -1,4 +1,0 @@
-
-Remove-Item .\src\.vs -Force -Recurse -ErrorAction Ignore
-Remove-Item .\src\obj -Force -Recurse -ErrorAction Ignore
-Remove-Item .\src\bin\Debug\*.runtimeconfig.dev.json -Force -Recurse -ErrorAction Ignore