+++ src/Generic/AppHelp.cs
@@ -1,11 +1,27 @@
 ﻿using System.Reflection;
 using System.Linq;
 using Microsoft.Win32;
+using System;
+using System.Text;
 
 class AppHelp
 {
     public static string ProductName { get; } = (Assembly.GetEntryAssembly()?.GetCustomAttributes(
         typeof(AssemblyProductAttribute)).First() as AssemblyProductAttribute)?.Product ?? "";
+
+    static string _ExecutablePath = "";
+
+    public static string ExecutablePath {
+        get {
+            if (_ExecutablePath == "")
+            {
+                StringBuilder sb = new StringBuilder(260);
+                TaskDialogNative.GetModuleFileName(IntPtr.Zero, sb, sb.Capacity);
+                _ExecutablePath = sb.ToString();
+            }
+            return _ExecutablePath;
+        }
+    }
 
     public static bool IsDarkTheme {
         get {